Updated `GameStartResponse.UtcTime` return type

+++ Core/Models/Eft/Game/GameStartResponse.cs
@@ -5,5 +5,5 @@ namespace Core.Models.Eft.Game;
 public class GameStartResponse
 {
     [JsonPropertyName("utc_time")]
-    public double? UtcTime { get; set; }
+    public long? UtcTime { get; set; }
 }
